requirements.txt
requests>=2.28.0
pyserial>=3.5
paho-mqtt>=1.6.0
